Accounting for 'Featuring' and apostrophies
Billboard 100 scraped artist data will include "Featuring", or other alternatives, and then said other artist.  Spotify, and Spotipy be extension, needed the artist to be one artist, or for the multiple artists to be separated by " AND ", for the query. Added code to account for this. Also added code to account for apostrophes in song names.